buck2_forkserver_proto

Struct CommandRequest

pub struct CommandRequest {
    pub exe: Vec<u8>,
    pub argv: Vec<Vec<u8>>,
    pub cwd: Option<WorkingDirectory>,
    pub timeout: Option<Duration>,
    pub env: Vec<EnvDirective>,
    pub enable_miniperf: bool,
    pub std_redirects: Option<StdRedirectPaths>,
    pub graceful_shutdown_timeout_s: Option<u32>,
    pub action_digest: Option<String>,
}

Fields
exe: Vec<u8>

The executable that should be run.

argv: Vec<Vec<u8>>

The arguments to give the executable.

cwd: Option<WorkingDirectory>

The working directory of the command as itโs being spawned.

timeout: Option<Duration>

The max duration this command should be allowed to run for.

env: Vec<EnvDirective>

Control the environment

enable_miniperf: bool

Enable Miniperf if available?

std_redirects: Option<StdRedirectPaths>

Used to optionally redirect stdout and stderr to files. If set, stdout and stderr events will not be streamed.

graceful_shutdown_timeout_s: Option<u32>

If set, how long to give the process time to shutdown after SIGTERM before sending SIGKILL. Should only be needed for daemonized processes (workers).

action_digest: Option<String>

Action digest is used when run actions through systemd, as we use it to create an unique cgroup name for action

Implementations

impl CommandRequest

pub fn graceful_shutdown_timeout_s(&self) -> u32

Returns the value of graceful_shutdown_timeout_s, or the default value if graceful_shutdown_timeout_s is unset.

pub fn action_digest(&self) -> &str

Returns the value of action_digest, or the default value if action_digest is unset.